fix radiator control
The latest state of the climate domain is more consistent with respect
to setting a temperature and switching the radiator off.

The former workaround to only send the hvac mode for certain HW is
removed as the mode is always sent when setting a target temperature.

Switching off is now done through a dedicated service.